+++ src/games/stendhal/server/maps/quests/revivalweeks/PhotographerChatAction.java
@@ -75,7 +75,7 @@ public class PhotographerChatAction implements ChatAction {
 
 	@Override
 	public void fire(Player player, Sentence sentence, EventRaiser npc) {
-		String outfit = Integer.toString(player.getOutfit().getCode());
+		String outfit = Long.toString(player.getOutfit().getCode());
 		int i = determinePhoto(player);
 
 		String url = generateUrl(outfit, i);
